ViewModel LikeFragment
